+++ addOns/pscanrules/src/main/java/org/zaproxy/zap/extension/pscanrules/ViewstateScanRule.java
@@ -86,7 +86,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 .setOtherInfo(var.getResultExtract().toString())
                 .setSolution(getSolution())
                 .setCweId(getCweId())
-                .setWascId(getWascId());
+                .setWascId(getWascId())
+                .setAlertRef(PLUGIN_ID + "-" + var.getAlertRef());
     }
 
     private AlertBuilder alertOldAspVersion() {
@@ -97,7 +98,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 .setDescription(Constant.messages.getString(MESSAGE_PREFIX + "oldver.desc"))
                 .setSolution(Constant.messages.getString(MESSAGE_PREFIX + "oldver.soln"))
                 .setCweId(getCweId())
-                .setWascId(getWascId());
+                .setWascId(getWascId())
+                .setAlertRef(PLUGIN_ID + "-3");
     }
 
     // TODO: see if this alert triggers too often, as the detection rule is far from being robust
@@ -111,7 +113,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 .setSolution(Constant.messages.getString(MESSAGE_PREFIX + "nomac.unsure.soln"))
                 .setReference(Constant.messages.getString(MESSAGE_PREFIX + "nomac.unsure.refs"))
                 .setCweId(getCweId())
-                .setWascId(getWascId());
+                .setWascId(getWascId())
+                .setAlertRef(PLUGIN_ID + "-4");
     }
 
     private AlertBuilder alertNoMACforSure() {
@@ -123,7 +126,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 .setSolution(Constant.messages.getString(MESSAGE_PREFIX + "nomac.sure.soln"))
                 .setReference(Constant.messages.getString(MESSAGE_PREFIX + "nomac.sure.refs"))
                 .setCweId(getCweId())
-                .setWascId(getWascId());
+                .setWascId(getWascId())
+                .setAlertRef(PLUGIN_ID + "-5");
     }
 
     private AlertBuilder alertSplitViewstate() {
@@ -134,7 +138,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 .setDescription(Constant.messages.getString(MESSAGE_PREFIX + "split.desc"))
                 .setSolution(Constant.messages.getString(MESSAGE_PREFIX + "split.soln"))
                 .setCweId(getCweId())
-                .setWascId(getWascId());
+                .setWascId(getWascId())
+                .setAlertRef(PLUGIN_ID + "-6");
     }
 
     @Override
@@ -256,6 +261,10 @@ public class ViewstateScanRule extends PluginPassiveScanner {
         public boolean hasResults() {
             return !this.resultExtract.isEmpty();
         }
+
+        public int getAlertRef() {
+            return this.pattern.getAlertRef();
+        }
     }
 
     // TODO: enhance this class with searches for e.g. passwords, ODBC strings, etc
@@ -265,7 +274,8 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                         "[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,4}", Pattern.CASE_INSENSITIVE),
                 Constant.messages.getString(MESSAGE_PREFIX + "content.email.name"),
                 Constant.messages.getString(MESSAGE_PREFIX + "content.email.desc"),
-                Constant.messages.getString(MESSAGE_PREFIX + "content.email.pattern.source")),
+                Constant.messages.getString(MESSAGE_PREFIX + "content.email.pattern.source"),
+                2),
 
         // TODO: once the viewstate parser is implemented, filter out all the version numbers of the
         // serialized objects which also trigger this filter
@@ -276,20 +286,27 @@ public class ViewstateScanRule extends PluginPassiveScanner {
                 Pattern.compile("\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}"),
                 Constant.messages.getString(MESSAGE_PREFIX + "content.ip.name"),
                 Constant.messages.getString(MESSAGE_PREFIX + "content.ip.desc"),
-                Constant.messages.getString(MESSAGE_PREFIX + "content.ip.pattern.source"));
+                Constant.messages.getString(MESSAGE_PREFIX + "content.ip.pattern.source"),
+                1);
 
         ViewstateAnalyzerPattern(
-                Pattern p, String alertHeader, String alertDescription, String sourceRegex) {
+                Pattern p,
+                String alertHeader,
+                String alertDescription,
+                String sourceRegex,
+                int alertRef) {
             this.pattern = p;
             this.alertHeader = alertHeader;
             this.alertDescription = alertDescription;
             this.sourceRegex = sourceRegex;
+            this.alertRef = alertRef;
         }
 
         private Pattern pattern;
         private String alertHeader;
         private String alertDescription;
         private String sourceRegex;
+        private final int alertRef;
 
         public Pattern getPattern() {
             return this.pattern;
@@ -301,6 +318,10 @@ public class ViewstateScanRule extends PluginPassiveScanner {
 
         public String getAlertHeader() {
             return this.alertHeader;
+        }
+
+        public int getAlertRef() {
+            return this.alertRef;
         }
     }